@@ -1,10 +1,13 @@
 {
     "name": "grimzy/laravel-mysql-spatial",
     "description": "MySQL spatial data types extension for Laravel.",
+    "homepage": "https://github.com/grimzy/laravel-mysql-spatial",
     "scripts": {
-        "test": "phpunit -c phpunit.xml.dist",
-        "test:unit": "phpunit -c phpunit.xml.dist --testsuite unit",
-        "test:integration": "phpunit -c phpunit.xml.dist --testsuite integration"
+        "post-autoload-dump": "@php ./vendor/bin/testbench package:discover --ansi",
+        "analyse": "vendor/bin/phpstan analyse",
+        "test": "vendor/bin/phpunit",
+        "test-coverage": "vendor/bin/phpunit --coverage",
+        "format": "vendor/bin/pint"
     },
     "type": "library",
     "license": "MIT",
@@ -15,39 +18,53 @@
         }
     ],
     "require": {
-        "php": ">=7.3",
-        "ext-pdo": "*",
+        "php": "^8.1",
         "ext-json": "*",
-        "illuminate/database": "^8.0",
+        "ext-pdo": "*",
+        "doctrine/dbal": "^3.5",
         "geo-io/wkb-parser": "^1.0",
+        "illuminate/contracts": "^10.0",
+        "illuminate/database": "^10.0",
         "jmikola/geojson": "^1.0"
     },
+    "conflict": {},
     "require-dev": {
-        "phpunit/phpunit": "~6.5",
-        "laravel/laravel": "^8.0",
-        "doctrine/dbal": "^2.5",
-        "laravel/browser-kit-testing": "^2.0",
-        "mockery/mockery": "^1.3"
+        "laravel/pint": "^1.6",
+        "laravel/laravel": "^10.0.4",
+        "laravel/browser-kit-testing": "^7.0",
+        "mockery/mockery": "^1.5.1",
+        "nunomaduro/collision": "^7.1.0",
+        "nunomaduro/larastan": "^2.5.1",
+        "orchestra/testbench": "^8.0.8",
+        "phpstan/extension-installer": "^1.2",
+        "phpstan/phpstan-deprecation-rules": "^1.1.2",
+        "phpstan/phpstan-phpunit": "^1.3.10",
+        "phpunit/phpunit": "^10.0.15"
     },
     "autoload": {
         "psr-4": {
-            "Grimzy\\LaravelMysqlSpatial\\": "src/"
+            "Grimzy\\LaravelMysqlSpatial\\": "src"
         }
     },
-    "autoload-dev" : {
-        "classmap" : [
-            "tests/Unit",
-            "tests/Integration"
-        ]
+    "autoload-dev": {
+        "psr-4": {
+            "Grimzy\\LaravelMysqlSpatial\\Tests\\": "tests"
+        }
+    },
+    "config": {
+        "sort-packages": true,
+        "allow-plugins": {
+            "pestphp/pest-plugin": true,
+            "phpstan/extension-installer": true
+        }
     },
     "extra": {
-        "branch-alias": {
-            "dev-master": "4.0.x-dev"
-        },
         "laravel": {
             "providers": [
                 "Grimzy\\LaravelMysqlSpatial\\SpatialServiceProvider"
             ]
         }
-    }
+    },
+    "minimum-stability": "stable",
+    "prefer-stable": true
 }
