@@ -3,13 +3,14 @@
 namespace Grimzy\LaravelMysqlSpatial\Types;
 
 use GeoIO\WKB\Parser\Parser;
+use GeoJson\Feature\Feature;
 use GeoJson\GeoJson;
 use Grimzy\LaravelMysqlSpatial\Exceptions\UnknownWKTTypeException;
 use Illuminate\Contracts\Support\Jsonable;
 
 abstract class Geometry implements GeometryInterface, Jsonable, \JsonSerializable
 {
-    protected static $wkb_types = [
+    protected static array$wkb_types = [
         1 => Point::class,
         2 => LineString::class,
         3 => Polygon::class,
@@ -19,24 +20,24 @@ abstract class Geometry implements GeometryInterface, Jsonable, \JsonSerializabl
         7 => GeometryCollection::class,
     ];
 
-    protected $srid;
+    protected int $srid;
 
-    public function __construct($srid = 0)
+    public function __construct(int $srid = 0)
     {
-        $this->srid = (int) $srid;
+        $this->srid = $srid;
     }
 
-    public function getSrid()
+    public function getSrid(): int
     {
         return $this->srid;
     }
 
-    public function setSrid($srid)
+    public function setSrid(int $srid): void
     {
-        $this->srid = (int) $srid;
+        $this->srid = $srid;
     }
 
-    public static function getWKTArgument($value)
+    public static function getWKTArgument(string $value): string
     {
         $left = strpos($value, '(');
         $right = strrpos($value, ')');
@@ -44,32 +45,25 @@ abstract class Geometry implements GeometryInterface, Jsonable, \JsonSerializabl
         return substr($value, $left + 1, $right - $left - 1);
     }
 
-    public static function getWKTClass($value)
+    /** @return class-string */
+    public static function getWKTClass(string $value): string
     {
         $left = strpos($value, '(');
         $type = trim(substr($value, 0, $left));
 
-        switch (strtoupper($type)) {
-            case 'POINT':
-                return Point::class;
-            case 'LINESTRING':
-                return LineString::class;
-            case 'POLYGON':
-                return Polygon::class;
-            case 'MULTIPOINT':
-                return MultiPoint::class;
-            case 'MULTILINESTRING':
-                return MultiLineString::class;
-            case 'MULTIPOLYGON':
-                return MultiPolygon::class;
-            case 'GEOMETRYCOLLECTION':
-                return GeometryCollection::class;
-            default:
-                throw new UnknownWKTTypeException('Type was '.$type);
-        }
+        return match (strtoupper($type)) {
+            'POINT' => Point::class,
+            'LINESTRING' => LineString::class,
+            'POLYGON' => Polygon::class,
+            'MULTIPOINT' => MultiPoint::class,
+            'MULTILINESTRING' => MultiLineString::class,
+            'MULTIPOLYGON' => MultiPolygon::class,
+            'GEOMETRYCOLLECTION' => GeometryCollection::class,
+            default => throw new UnknownWKTTypeException('Type was '.$type)
+        };
     }
 
-    public static function fromWKB($wkb)
+    public static function fromWKB(string $wkb): Geometry
     {
         $srid = substr($wkb, 0, 4);
         $srid = unpack('L', $srid)[1];
@@ -87,14 +81,14 @@ abstract class Geometry implements GeometryInterface, Jsonable, \JsonSerializabl
         return $parsed;
     }
 
-    public static function fromWKT($wkt, $srid = null)
+    public static function fromWKT(string $wkt, int $srid = 0): static
     {
         $wktArgument = static::getWKTArgument($wkt);
 
         return static::fromString($wktArgument, $srid);
     }
 
-    public static function fromJson($geoJson)
+    public static function fromJson(string|GeoJson $geoJson): self
     {
         if (is_string($geoJson)) {
             $geoJson = GeoJson::jsonUnserialize(json_decode($geoJson));
@@ -104,6 +98,7 @@ abstract class Geometry implements GeometryInterface, Jsonable, \JsonSerializabl
             return GeometryCollection::fromJson($geoJson);
         }
 
+        /** @var Feature $geoJson */
         if ($geoJson->getType() === 'Feature') {
             $geoJson = $geoJson->getGeometry();
         }
