@@ -6,63 +6,60 @@ use GeoJson\GeoJson;
 use GeoJson\Geometry\MultiLineString as GeoJsonMultiLineString;
 use Grimzy\LaravelMysqlSpatial\Exceptions\InvalidGeoJsonException;
 
-class MultiLineString extends GeometryCollection
+/**
+ * @implements GeometryInterface<LineString>
+ *
+ * @extends GeometryCollection<LineString>
+ */
+class MultiLineString extends GeometryCollection implements GeometryInterface
 {
     /**
      * The minimum number of items required to create this collection.
-     *
-     * @var int
      */
-    protected $minimumCollectionItems = 1;
+    protected int $minimumCollectionItems = 1;
 
     /**
      * The class of the items in the collection.
-     *
-     * @var string
      */
-    protected $collectionItemType = LineString::class;
+    protected string $collectionItemType = LineString::class;
 
-    public function getLineStrings()
+    public function getLineStrings(): array
     {
         return $this->items;
     }
 
-    public function toWKT()
+    public function toWKT(): string
     {
         return sprintf('MULTILINESTRING(%s)', (string) $this);
     }
 
-    public static function fromString($wktArgument, $srid = 0)
+    public static function fromString(string $wktArgument, int $srid = 0): static
     {
         $str = preg_split('/\)\s*,\s*\(/', substr(trim($wktArgument), 1, -1));
-        $lineStrings = array_map(function ($data) {
-            return LineString::fromString($data);
-        }, $str);
+        $lineStrings = array_map(fn ($data) => LineString::fromString($data), $str);
 
         return new static($lineStrings, $srid);
     }
 
     public function __toString()
     {
-        return implode(',', array_map(function (LineString $lineString) {
-            return sprintf('(%s)', (string) $lineString);
-        }, $this->getLineStrings()));
+        return implode(',', array_map(fn (LineString $lineString) => sprintf('(%s)', (string) $lineString), $this->getLineStrings()));
     }
 
-    public function offsetSet($offset, $value)
+    public function offsetSet($offset, $value): void
     {
         $this->validateItemType($value);
 
         parent::offsetSet($offset, $value);
     }
 
-    public static function fromJson($geoJson)
+    public static function fromJson(string|GeoJson $geoJson): self
     {
         if (is_string($geoJson)) {
             $geoJson = GeoJson::jsonUnserialize(json_decode($geoJson));
         }
 
-        if (!is_a($geoJson, GeoJsonMultiLineString::class)) {
+        if (! is_a($geoJson, GeoJsonMultiLineString::class)) {
             throw new InvalidGeoJsonException('Expected '.GeoJsonMultiLineString::class.', got '.get_class($geoJson));
         }
 
@@ -80,9 +77,8 @@ class MultiLineString extends GeometryCollection
 
     /**
      * Convert to GeoJson Point that is jsonable to GeoJSON.
-     *
-     * @return \GeoJson\Geometry\MultiLineString
      */
+    #[\ReturnTypeWillChange]
     public function jsonSerialize()
     {
         $lineStrings = [];
