@@ -1,20 +1,17 @@
 <?php
 
-namespace Schema;
+namespace Grimzy\LaravelMysqlSpatial\Tests\Unit\Schema;
 
-use BaseTestCase;
 use Grimzy\LaravelMysqlSpatial\Schema\Blueprint;
+use Grimzy\LaravelMysqlSpatial\Tests\Unit\BaseTestCase as UnitBaseTestCase;
 use Illuminate\Database\Schema\ColumnDefinition;
 use Mockery;
 
-class BlueprintTest extends BaseTestCase
+class BlueprintTest extends UnitBaseTestCase
 {
-    /**
-     * @var \Grimzy\LaravelMysqlSpatial\Schema\Blueprint
-     */
-    protected $blueprint;
+    protected \Grimzy\LaravelMysqlSpatial\Schema\Blueprint $blueprint;
 
-    public function setUp()
+    public function setUp(): void
     {
         parent::setUp();
 
