@@ -9,12 +9,10 @@ use PDO;
 class ConnectionFactory extends IlluminateConnectionFactory
 {
     /**
-     * @param string       $driver
-     * @param \Closure|PDO $connection
-     * @param string       $database
-     * @param string       $prefix
-     * @param array        $config
-     *
+     * @param  string  $driver
+     * @param  \Closure|PDO  $connection
+     * @param  string  $database
+     * @param  string  $prefix
      * @return \Illuminate\Database\ConnectionInterface
      */
     protected function createConnection($driver, $connection, $database, $prefix = '', array $config = [])
