@@ -1,20 +1,23 @@
 <?php
 
+namespace Grimzy\LaravelMysqlSpatial\Tests\Integration;
+
 use Grimzy\LaravelMysqlSpatial\SpatialServiceProvider;
+use Grimzy\LaravelMysqlSpatial\Tests\Integration\Migrations\CreateTables;
+use Grimzy\LaravelMysqlSpatial\Tests\Integration\Migrations\UpdateTables;
 use Illuminate\Support\Facades\DB;
 use Laravel\BrowserKitTesting\TestCase as BaseTestCase;
 
-abstract class IntegrationBaseTestCase extends BaseTestCase
+abstract class IntegrationBaseCase extends BaseTestCase
 {
     protected $after_fix = false;
+
     protected $migrations = [];
 
     /**
      * Boots the application.
-     *
-     * @return \Illuminate\Foundation\Application
      */
-    public function createApplication()
+    public function createApplication(): \Illuminate\Foundation\Application
     {
         $app = require __DIR__.'/../../vendor/laravel/laravel/bootstrap/app.php';
         $app->register(SpatialServiceProvider::class);
@@ -41,29 +44,27 @@ abstract class IntegrationBaseTestCase extends BaseTestCase
 
     /**
      * Setup DB before each test.
-     *
-     * @return void
      */
-    public function setUp()
+    public function setUp(): void
     {
         parent::setUp();
 
         $this->after_fix = $this->isMySQL8AfterFix();
 
-        $this->onMigrations(function ($migrationClass) {
-            (new $migrationClass())->up();
-        });
+        $this->artisan('migrate:fresh');
+
+        (new CreateTables)->up();
+        (new UpdateTables)->up();
 
         //\DB::listen(function($sql) {
         //    var_dump($sql);
         //});
     }
 
-    public function tearDown()
+    public function tearDown(): void
     {
-        $this->onMigrations(function ($migrationClass) {
-            (new $migrationClass())->down();
-        }, true);
+        (new UpdateTables)->down();
+        (new CreateTables)->down();
 
         parent::tearDown();
     }
@@ -71,7 +72,7 @@ abstract class IntegrationBaseTestCase extends BaseTestCase
     // MySQL 8.0.4 fixed bug #26941370 and bug #88031
     private function isMySQL8AfterFix()
     {
-        $results = DB::select(DB::raw('select version()'));
+        $results = DB::select(DB::raw('select version()')->getValue(DB::connection()->getQueryGrammar()));
         $mysql_version = $results[0]->{'version()'};
 
         return version_compare($mysql_version, '8.0.4', '>=');
@@ -90,7 +91,7 @@ abstract class IntegrationBaseTestCase extends BaseTestCase
     {
         if (method_exists(parent::class, 'expectException')) {
             parent::expectException($exceptionName);
-            if (!is_null($exceptionMessage)) {
+            if (! is_null($exceptionMessage)) {
                 $this->expectExceptionMessage($exceptionMessage);
             }
         } else {
@@ -98,7 +99,7 @@ abstract class IntegrationBaseTestCase extends BaseTestCase
         }
     }
 
-    private function onMigrations(\Closure $closure, $reverse_sort = false)
+    private function onMigrations(Closure $closure, $reverse_sort = false)
     {
         $migrations = $this->migrations;
         $reverse_sort ? rsort($migrations, SORT_STRING) : sort($migrations, SORT_STRING);
