@@ -1,18 +1,32 @@
 <?php
 
+namespace Grimzy\LaravelMysqlSpatial\Tests\Integration\Migrations;
+
 use Grimzy\LaravelMysqlSpatial\Schema\Blueprint;
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Support\Facades\Schema;
 
-class CreateLocationTable extends Migration
+class CreateTables extends Migration
 {
     /**
      * Run the migrations.
-     *
-     * @return void
      */
-    public function up()
+    public function up(): void
     {
+        Schema::create('test_models', function (Blueprint $table) {
+            $table->charset = 'utf8mb4';
+            $table->collation = 'utf8mb4_unicode_ci';
+            $table->increments('id');
+            $table->geometryCollection('geometrycollection')->default(null)->nullable();
+            $table->lineString('linestring')->default(null)->nullable();
+            $table->multiLineString('multilinestring')->default(null)->nullable();
+            $table->multiPoint('multipoint')->default(null)->nullable();
+            $table->multiPolygon('multipolygon')->default(null)->nullable();
+            $table->point('point')->default(null)->nullable();
+            $table->polygon('polygon')->default(null)->nullable();
+            $table->timestamps();
+        });
+
         Schema::create('geometry', function (Blueprint $table) {
             $table->charset = 'utf8mb4';
             $table->collation = 'utf8mb4_unicode_ci';
@@ -50,13 +64,12 @@ class CreateLocationTable extends Migration
 
     /**
      * Reverse the migrations.
-     *
-     * @return void
      */
-    public function down()
+    public function down(): void
     {
         Schema::drop('geometry');
         Schema::drop('no_spatial_fields');
         Schema::drop('with_srid');
+        Schema::drop('test_models');
     }
 }
