@@ -1,14 +1,14 @@
 <?php
 
-namespace Schema;
+namespace Grimzy\LaravelMysqlSpatial\Tests\Unit\Schema;
 
-use BaseTestCase;
 use Grimzy\LaravelMysqlSpatial\MysqlConnection;
 use Grimzy\LaravelMysqlSpatial\Schema\Blueprint;
 use Grimzy\LaravelMysqlSpatial\Schema\Builder;
+use Grimzy\LaravelMysqlSpatial\Tests\Unit\BaseTestCase as UnitBaseTestCase;
 use Mockery;
 
-class BuilderTest extends BaseTestCase
+class BuilderTest extends UnitBaseTestCase
 {
     public function testReturnsCorrectBlueprint()
     {
