@@ -1,10 +1,14 @@
 <?php
 
+namespace Grimzy\LaravelMysqlSpatial\Tests\Unit\Types;
+
+use Grimzy\LaravelMysqlSpatial\Tests\Unit\BaseTestCase;
 use Grimzy\LaravelMysqlSpatial\Types\GeometryCollection;
 use Grimzy\LaravelMysqlSpatial\Types\GeometryInterface;
 use Grimzy\LaravelMysqlSpatial\Types\LineString;
 use Grimzy\LaravelMysqlSpatial\Types\Point;
 use Grimzy\LaravelMysqlSpatial\Types\Polygon;
+use InvalidArgumentException;
 
 class GeometryCollectionTest extends BaseTestCase
 {
@@ -71,7 +75,7 @@ class GeometryCollectionTest extends BaseTestCase
     {
         $geometryCollection = $this->getGeometryCollection();
 
-        $this->assertInternalType('array', $geometryCollection->toArray());
+        $this->assertIsArray($geometryCollection->toArray());
     }
 
     public function testIteratorAggregate()
@@ -131,7 +135,7 @@ class GeometryCollectionTest extends BaseTestCase
     {
         $this->assertException(
             \Grimzy\LaravelMysqlSpatial\Exceptions\InvalidGeoJsonException::class,
-            sprintf('Expected %s, got %s', GeoJson\Feature\FeatureCollection::class, GeoJson\Geometry\Point::class)
+            sprintf('Expected %s, got %s', \GeoJson\Feature\FeatureCollection::class, \GeoJson\Geometry\Point::class)
         );
         GeometryCollection::fromJson('{"type":"Point","coordinates":[3.4,1.2]}');
     }
